requirements.txt
streammachine-driver==1.0.0
streammachine-schemas-demo-avro==1.0.1
